Crate fusiontables2

This documentation was generated from fusiontables crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

column (delete, get, insert, list, patch and update)

query (sql and sql get)

style (delete, get, insert, list, patch and update)

table (copy, delete, get, import rows, import table, insert, list, patch, replace rows and update)

task (delete, get and list)

template (delete, get, insert, list, patch and update)

Everything else about the fusiontables v2 API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.table().import_table(...).doit()
let r = hub.table().patch(...).doit()
let r = hub.table().get(...).doit()
let r = hub.table().copy(...).doit()
let r = hub.table().insert(...).doit()
let r = hub.table().update(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
fusiontables2 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate fusiontables2;
use fusiontables2::cmn::Result;
use std::fs;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use fusiontables2::Fusiontables;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Fusiontables::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `upload_resumable(...)`.
// Values shown here are possibly random and not representative !
let result = hub.table().import_table("name")
             .encoding("et")
             .delimiter("dolores")
             .upload_resumable(fs::File::open("file.ext").unwrap(), 282, "application/octet-stream".parse().unwrap());
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

Bucket
Specifies the minimum and maximum values, the color, opacity, icon and weight of a bucket within a StyleSetting.

This type is not used in any activity, and only used as part of another schema.

Column
Specifies the details of a column in a table.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get column (response)

update column (request|response)

insert column (request|response)

patch column (request|response)

ColumnBaseColumn
Identifier of the base column. If present, this column is derived from the specified base column.

This type is not used in any activity, and only used as part of another schema.

ColumnDeleteMethodBuilder
Deletes the column.

ColumnGetMethodBuilder
Retrieves a specific column by its id.

ColumnInsertMethodBuilder
Adds a new column to the table.

ColumnList
Represents a list of columns in a table.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list column (response)

ColumnListMethodBuilder
Retrieves a list of columns.

ColumnMethodsBuilder
A builder providing access to all methods supported on column resources. It is not used directly, but through the Fusiontables hub.

ColumnPatchMethodBuilder
Updates the name or type of an existing column. This method supports patch semantics.

ColumnUpdateMethodBuilder
Updates the name or type of an existing column.

Fusiontables
Central instance to access all Fusiontables related resource activities

Geometry
Represents a Geometry object.

This type is not used in any activity, and only used as part of another schema.

Import
Represents an import request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

import rows table (response)

Line
Represents a line geometry.

This type is not used in any activity, and only used as part of another schema.

LineStyle
Represents a LineStyle within a StyleSetting

This type is not used in any activity, and only used as part of another schema.

Point
Represents a point object.

This type is not used in any activity, and only used as part of another schema.

PointStyle
Represents a PointStyle within a StyleSetting

This type is not used in any activity, and only used as part of another schema.

Polygon
Represents a polygon object.

This type is not used in any activity, and only used as part of another schema.

PolygonStyle
Represents a PolygonStyle within a StyleSetting

This type is not used in any activity, and only used as part of another schema.

QueryMethodsBuilder
A builder providing access to all methods supported on query resources. It is not used directly, but through the Fusiontables hub.

QuerySqlGetMethodBuilder
Executes an SQL SELECT/SHOW/DESCRIBE statement.

QuerySqlMethodBuilder
Executes an SQL SELECT/INSERT/UPDATE/DELETE/SHOW/DESCRIBE/CREATE statement.

Sqlresponse
Represents a response to an sql statement.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

sql get query (response)

sql query (response)

StyleDeleteMethodBuilder
Deletes a style.

StyleFunction
Represents a StyleFunction within a StyleSetting

This type is not used in any activity, and only used as part of another schema.

StyleFunctionGradient
Gradient function that interpolates a range of colors based on column value.

This type is not used in any activity, and only used as part of another schema.

StyleFunctionGradientColors
Array with two or more colors.

This type is not used in any activity, and only used as part of another schema.

StyleGetMethodBuilder
Gets a specific style.

StyleInsertMethodBuilder
Adds a new style for the table.

StyleListMethodBuilder
Retrieves a list of styles.

StyleMethodsBuilder
A builder providing access to all methods supported on style resources. It is not used directly, but through the Fusiontables hub.

StylePatchMethodBuilder
Updates an existing style. This method supports patch semantics.

StyleSetting
Represents a complete StyleSettings object. The primary key is a combination of the tableId and a styleId.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

patch style (request|response)

get style (response)

update style (request|response)

insert style (request|response)

StyleSettingList
Represents a list of styles for a given table.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list style (response)

StyleUpdateMethodBuilder
Updates an existing style.

Table
Represents a table.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

import table table (response)

patch table (request|response)

get table (response)

copy table (response)

insert table (request|response)

update table (request|response)

TableCopyMethodBuilder
Copies a table.

TableDeleteMethodBuilder
Deletes a table.

TableGetMethodBuilder
Retrieves a specific table by its id.

TableImportRowMethodBuilder
Import more rows into a table.

TableImportTableMethodBuilder
Import a new table.

TableInsertMethodBuilder
Creates a new table.

TableList
Represents a list of tables.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list table (response)

TableListMethodBuilder
Retrieves a list of tables a user owns.

TableMethodsBuilder
A builder providing access to all methods supported on table resources. It is not used directly, but through the Fusiontables hub.

TablePatchMethodBuilder
Updates an existing table. Unless explicitly requested, only the name, description, and attribution will be updated. This method supports patch semantics.

TableReplaceRowMethodBuilder
Replaces rows of an existing table. Current rows remain visible until all replacement rows are ready.

TableUpdateMethodBuilder
Updates an existing table. Unless explicitly requested, only the name, description, and attribution will be updated.

Task
A background task on a table, initiated for time and or resource consuming operations such as a column type change or delete all rows operation.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get task (response)

replace rows table (response)

TaskDeleteMethodBuilder
Deletes the task, unless already started.

TaskGetMethodBuilder
Retrieves a specific task by its id.

TaskList
Represents a list of tasks in a table.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list task (response)

TaskListMethodBuilder
Retrieves a list of tasks.

TaskMethodsBuilder
A builder providing access to all methods supported on task resources. It is not used directly, but through the Fusiontables hub.

Template
Represents the contents of InfoWindow templates.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get template (response)

update template (request|response)

insert template (request|response)

patch template (request|response)

TemplateDeleteMethodBuilder
Deletes a template

TemplateGetMethodBuilder
Retrieves a specific template by its id

TemplateInsertMethodBuilder
Creates a new template for the table.

TemplateList
Represents a list of templates for a given table.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list template (response)

TemplateListMethodBuilder
Retrieves a list of templates.

TemplateMethodsBuilder
A builder providing access to all methods supported on template resources. It is not used directly, but through the Fusiontables hub.

TemplatePatchMethodBuilder
Updates an existing template. This method supports patch semantics.

TemplateUpdateMethodBuilder
Updates an existing template

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.