Struct fusiontables2::Table

pub struct Table {
    pub table_properties_json_schema: Option<String>,
    pub kind: Option<String>,
    pub attribution: Option<String>,
    pub description: Option<String>,
    pub column_properties_json_schema: Option<String>,
    pub is_exportable: Option<bool>,
    pub base_table_ids: Vec<String>,
    pub table_properties_json: Option<String>,
    pub attribution_link: Option<String>,
    pub sql: Option<String>,
    pub table_id: Option<String>,
    pub columns: Vec<Column>,
    pub name: Option<String>,
}

Represents a table.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

importTable (response)

patch (request|response)

get (response)

copy (response)

insert (request|response)

update (request|response)

Fields

table_properties_json_schema

JSON schema for validating the JSON table properties.

kind

Type name: a template for an individual table.

attribution

Attribution assigned to the table.

description

Description assigned to the table.

column_properties_json_schema

Default JSON schema for validating all JSON column properties.

is_exportable

Variable for whether table is exportable.

base_table_ids

Base table identifier if this table is a view or merged table.

table_properties_json

JSON object containing custom table properties.

attribution_link

Optional link for attribution.

sql

SQL that encodes the table definition for derived tables.

table_id

Encrypted unique alphanumeric identifier for the table.

columns

Columns in the table.

name

Name assigned to a table.