Struct fusiontables2::TableReplaceRowMethodBuilder

pub struct TableReplaceRowMethodBuilder<'a, C, NC, A> where NC: 'a, C: 'a, A: 'a {
    // some fields omitted
}

Replaces rows of an existing table. Current rows remain visible until all replacement rows are ready.

A builder for the replaceRows method supported by a table resource. It is not used directly, but through a TableMethodsBuilder.

Example

Instantiate a resource method builder

use std::fs;
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `upload_resumable(...)`.
// Values shown here are possibly random and not representative !
let result = hub.table().replace_rows("tableId")
             .start_line(-48)
             .is_strict(true)
             .end_line(-23)
             .encoding("consetetur")
             .delimiter("amet.")
             .upload_resumable(fs::File::open("file.ext").unwrap(), 282, "application/octet-stream".parse().unwrap());

Methods

impl<'a, C, NC, A> TableReplaceRowMethodBuilder<'a, C, NC, A> where NC: NetworkConnector, C: BorrowMut<Client<NC>> + 'a, A: GetToken

fn upload<R>(self, stream: R, size: u64, mime_type: Mime) -> Result<Task> where R: Read

Upload media all at once. If the upload fails for whichever reason, all progress is lost.

max size: 250MB

multipart: yes

valid mime types: 'application/octet-stream'

fn upload_resumable<RS>(self, resumeable_stream: RS, size: u64, mime_type: Mime) -> Result<Task> where RS: ReadSeek

Upload media in a resumeable fashion. Even if the upload fails or is interrupted, it can be resumed for a certain amount of time as the server maintains state temporarily.

TODO: Write more about how delegation works in this particular case.

max size: 250MB

multipart: yes

valid mime types: 'application/octet-stream'

fn table_id(self, new_value: &str) -> TableReplaceRowMethodBuilder<'a, C, NC, A>

Sets the table id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

Table whose rows will be replaced.

fn start_line(self, new_value: i32) -> TableReplaceRowMethodBuilder<'a, C, NC, A>

Sets the start line query property to the given value.

The index of the first line from which to start importing, inclusive. Default is 0.

fn is_strict(self, new_value: bool) -> TableReplaceRowMethodBuilder<'a, C, NC, A>

Sets the is strict query property to the given value.

Whether the CSV must have the same number of column values for each row. If true, throws an exception if the CSV does not not have the same number of columns. If false, rows with fewer column values will be padded with empty values. Default is true.

fn end_line(self, new_value: i32) -> TableReplaceRowMethodBuilder<'a, C, NC, A>

Sets the end line query property to the given value.

The index of the last line to import, exclusive. 'endLine - startLine' rows will be imported. Default is to import through the end of the file. If endLine is negative, it is an offset from the end of the file; the imported content will exclude the last endLine lines.

fn encoding(self, new_value: &str) -> TableReplaceRowMethodBuilder<'a, C, NC, A>

Sets the encoding query property to the given value.

The encoding of the content. Default is UTF-8. Use 'auto-detect' if you are unsure of the encoding.

fn delimiter(self, new_value: &str) -> TableReplaceRowMethodBuilder<'a, C, NC, A>

Sets the delimiter query property to the given value.

The delimiter used to separate cell values. This can only consist of a single character. Default is ','.

fn delegate(self, new_value: &'a mut Delegate) -> TableReplaceRowMethodBuilder<'a, C, NC, A>

Sets the delegate property to the given value.

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

fn param<T>(self, name: T, value: T) -> TableReplaceRowMethodBuilder<'a, C, NC, A> where T: Str

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> TableReplaceRowMethodBuilder<'a, C, NC, A> where T: Str

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead of relying on the automated algorithm which simply prefers read-only scopes over those who are not.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.